Add tool for managing string translations
Since we've got more than just the standard strings in resource files
and it would be beneficial to involve RST in the translation process,
I've cooked up a script to copy/paste strings to/from a Google Sheet.
This should make it easier to solicit strings from folks at RST, and
then add them back into the codebase.

resolves #57